+++ Dockerfile
@@ -1,0 +1,6 @@
+FROM python:3.7.4
+
+ARG project_directory
+WORKDIR $project_directory
+
+RUN pip install flask
