+++ docker-compose.yml
@@ -1,0 +1,15 @@
+version: "3.7"
+services:
+  web_flask:
+    build:
+      context: .
+      args:
+        project_directory: "/flaskr/"
+    ports:
+      - "5000:5000"
+    volumes:
+      - "./flaskr:/flaskr"
+    environment:
+      TZ: "Asia/Tokyo"
+      FLASK_ENV: "development"
+    command: "flask run --host 0.0.0.0 --port 5000"
